requirements.txt
asgiref==3.6.0
autopep8==2.0.1
black==22.12.0
click==8.1.3
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2022.7.1
sqlparse==0.4.3
